// src/api/models/bpmn-diagram-upload-to.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface BpmnDiagramUploadTO
 */
export interface BpmnDiagramUploadTO {
    /**
     * 
     * @type {string}
     * @memberof BpmnDiagramUploadTO
     */
    bpmnDiagramId?: any;
    /**
     * 
     * @type {string}
     * @memberof BpmnDiagramUploadTO
     */
    bpmnDiagramName: any;
    /**
     * 
     * @type {string}
     * @memberof BpmnDiagramUploadTO
     */
    bpmnDiagramDescription: any;
    /**
     * 
     * @type {string}
     * @memberof BpmnDiagramUploadTO
     */
    fileType?: any;
    /**
     * 
     * @type {string}
     * @memberof BpmnDiagramUploadTO
     */
    svgPreview?: any;
}